fix(build): add linuxstatic target and build output
The codecov documentation states that the uploader is a statically
compiled binary:
https://docs.codecov.com/docs/codecov-uploader

Unfortunately, the npm package -- pkg -- used to build
the static binaries generates a dynamically linked executable for linux
by default. To generate a static binary for linux, the linuxstatic
target has to be used.

I ran into this issue trying to use the uploader in a CI environment
where the necessary shared objects weren't all available.

Since it wasn't clear whether the dynamically linked executable was a
feature, I made the static build output an additional one.